Fix block updating in the custom blocks system and move the assets back to resource-pack/assets
