+++ run/server.properties
@@ -1,5 +1,5 @@
 #Minecraft server properties
-#Sun Mar 06 10:53:00 BRT 2022
+#Sun Mar 06 11:24:39 BRT 2022
 enable-jmx-monitoring=false
 rcon.port=25575
 level-seed=
